requirement.txt
langchain
plotly
pandas
streamlit
pdfplumber
datetime
python-dotenv
langchain-google-genai
langchain-pinecone
psycopg2